feat: Update xyzToLab to return string or object format
The xyzToLab function is now updated to return result in string or object format based on an additional optional boolean argument. This function now also validates the inputs and throws an error if any of the color values are out of range. Corresponding tests are updated to reflect these changes.