feat: add recomposeColor function and corresponding tests
This commit introduces a new function, recomposeColor, which is able to recompose color values into a color string based on the detected format. This feature comes with its own set of unit tests for ensuring a variety of color formats are correctly supported. Changes have been made on the index file to expose this new function to the modules consumers.