feat(parser): add LCH color string parsing functionality
This commit introduces a new parsing function `parseLch`, which can parse LCH color strings in the src/parser folder. A set of tests to verify the correct functioning of this parser are also added in tests/parser. The parseLch function will return an array containing the LCH components [L, C, H, A?], and will throw an error when encountering an invalid LCH color format.