feat: Rename package and update descriptions
This commit renames the package from 'colore' to 'colore-js' and updates all references to the package name in the documentation. Additionally, the description in the package.json file and the page title in the documentation index is updated to provide a more comprehensive overview of the library's capabilities. Also, 'tests/**/*' has been removed from the 'include' array in tsconfig.json, and several