feat: Update badges and README details
The color and label parameters for all the badges in README.md were removed, resulting in cleaner code. A small adjustment was made in the Color Conversion feature description for clarity. Additional features were added to the list, namely 'No Dependencies' and 'Small Size', giving a more comprehensive view of the library's attributes.