feat: add generateInterpolatedColors function
This new function interpolates between two colors of various formats (RGB, RGBA, HSL, HSLA, HEX, HEX_Alpha, LAB, LCH) in a specified number of steps. It can return the results as strings, or as objects. The function is now exported in the main index file and it has associated unit tests.