feat: Add detailed description and keywords to rgbToHex doc page
Added a comprehensive description and set of keywords for the rgbToHex function on the Colore JS documentation. This enhancement aims to give users a clearer understanding of the function's purpose and use by detailing its syntax, parameters, examples, and practical applications. Furthermore, newly added keywords enhance page's discoverability in search.